# Address ipcache startup perfomance regression