--- a/pkg/ipcache/ipcache_test.go
+++ b/pkg/ipcache/ipcache_test.go
@@ -5,10 +5,13 @@ package ipcache
 
 import (
 	"context"
+	"encoding/binary"
 	"errors"
 	"fmt"
 	"net"
+	"net/netip"
 	"sort"
+	"strconv"
 	"testing"
 
 	. "gopkg.in/check.v1"
@@ -551,6 +554,75 @@ func (s *IPCacheTestSuite) TestIPCacheNamedPorts(c *C) {
 	c.Assert(namedPortsChanged, Equals, true)
 	npm = IPIdentityCache.GetNamedPorts()
 	c.Assert(npm, HasLen, 0)
+}
+
+func BenchmarkIPCacheUpsert10(b *testing.B) {
+	benchmarkIPCacheUpsert(b, 10)
+}
+
+func BenchmarkIPCacheUpsert100(b *testing.B) {
+	benchmarkIPCacheUpsert(b, 100)
+}
+
+func BenchmarkIPCacheUpsert1000(b *testing.B) {
+	benchmarkIPCacheUpsert(b, 1000)
+}
+
+func BenchmarkIPCacheUpsert10000(b *testing.B) {
+	benchmarkIPCacheUpsert(b, 10000)
+}
+
+func benchmarkIPCacheUpsert(b *testing.B, num int) {
+	meta := K8sMetadata{
+		Namespace: "default",
+		PodName:   "app",
+		NamedPorts: types.NamedPortMap{
+			"http": types.PortProto{Port: 80, Proto: uint8(u8proto.TCP)},
+			"dns":  types.PortProto{Port: 53},
+		},
+	}
+
+	buf := make([]byte, 4)
+	ips := make([]string, num)
+	nms := make([]string, num)
+	for i := range nms {
+		binary.BigEndian.PutUint32(buf, uint32(i+2<<26))
+		ip, _ := netip.AddrFromSlice(buf)
+		ips[i] = ip.String()
+		nms[i] = strconv.Itoa(i)
+	}
+
+	b.StopTimer()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		ctx, cancel := context.WithCancel(context.Background())
+		allocator := testidentity.NewMockIdentityAllocator(nil)
+		ipcache := NewIPCache(&Configuration{
+			Context:           ctx,
+			IdentityAllocator: allocator,
+			PolicyHandler:     &mockUpdater{},
+			DatapathHandler:   &mockTriggerer{},
+			NodeHandler:       &mockNodeHandler{},
+		})
+
+		// We only want to measure the calls to upsert.
+		b.StartTimer()
+		for j := 0; j < num; j++ {
+			meta.PodName = nms[j]
+			_, err := ipcache.Upsert(ips[j], nil, 0, &meta, Identity{
+				ID:     identityPkg.NumericIdentity(j),
+				Source: source.Kubernetes,
+			})
+			if err != nil {
+				b.Fatalf("failed to upsert: %v", err)
+			}
+		}
+		b.StopTimer()
+
+		// Clean up after ourselves, so that the individual runs are comparable.
+		cancel()
+		ipcache.Shutdown()
+	}
 }
 
 type dummyListener struct {