--- a/pkg/types/portmap_test.go
+++ b/pkg/types/portmap_test.go
@@ -103,7 +103,7 @@ func (ds *PortsTestSuite) TestPolicyPortProtoSet(c *C) {
 }
 
 func (ds *PortsTestSuite) TestPolicyNamedPortMultiMap(c *C) {
-	a := NamedPortMultiMap{
+	a := namedPortMultiMap{
 		"http": PortProtoSet{
 			PortProto{Port: 80, Proto: 6}:   struct{}{},
 			PortProto{Port: 8080, Proto: 6}: struct{}{},
@@ -120,7 +120,7 @@ func (ds *PortsTestSuite) TestPolicyNamedPortMultiMap(c *C) {
 			PortProto{Port: 53, Proto: 6}:  struct{}{},
 		},
 	}
-	b := NamedPortMultiMap{
+	b := namedPortMultiMap{
 		"http": PortProtoSet{
 			PortProto{Port: 80, Proto: 6}:   struct{}{},
 			PortProto{Port: 8080, Proto: 6}: struct{}{},
@@ -161,7 +161,8 @@ func (ds *PortsTestSuite) TestPolicyNamedPortMultiMap(c *C) {
 	c.Assert(err, Equals, ErrUnknownNamedPort)
 	c.Assert(port, Equals, uint16(0))
 
-	port, err = NamedPortMultiMap(nil).GetNamedPort("unknown", 6)
+	var nilvalued namedPortMultiMap
+	port, err = NamedPortMultiMap(nilvalued).GetNamedPort("unknown", 6)
 	c.Assert(err, Equals, ErrNilMap)
 	c.Assert(port, Equals, uint16(0))