The commit messages should be descriptive, here's the summary line for each.

1. ipcache: add benchmark for Upsert (`regression.txt` below)
2. ipcache: make NamedPortMultiMap an interface (`interface.txt` below)
3. ipcache: switch named ports to reference counting (`fix.txt` below)
4. ipcache: simplify the named ports update code (`cleanup.txt` below)

The following is a comparison of running the benchmark added in commit 1 over the course of the four commits.

```
goos: linux
goarch: arm64
pkg: github.com/cilium/cilium/pkg/ipcache
                      │ regression.txt │             interface.txt             │               fix.txt               │             cleanup.txt             │
                      │     sec/op     │     sec/op      vs base               │   sec/op     vs base                │   sec/op     vs base                │
IPCacheUpsert10-10        97.10µ ±  3%     94.58µ ±  3%       ~ (p=0.052 n=10)   48.84µ ± 5%  -49.70% (p=0.000 n=10)   47.59µ ± 5%  -50.98% (p=0.000 n=10)
IPCacheUpsert100-10      2672.3µ ±  5%    2718.4µ ±  8%       ~ (p=0.481 n=10)   508.3µ ± 5%  -80.98% (p=0.000 n=10)   485.0µ ± 2%  -81.85% (p=0.000 n=10)
IPCacheUpsert1000-10     69.543m ± 13%    66.515m ± 17%       ~ (p=0.481 n=10)   5.282m ± 5%  -92.40% (p=0.000 n=10)   5.056m ± 2%  -92.73% (p=0.000 n=10)
IPCacheUpsert10000-10   4166.59m ±  1%   4299.95m ±  2%  +3.20% (p=0.000 n=10)   48.63m ± 5%  -98.83% (p=0.000 n=10)   44.87m ± 5%  -98.92% (p=0.000 n=10)
geomean                   16.56m           16.47m        -0.55%                  1.589m       -90.40%                  1.513m       -90.86%
```

The last commit, `cleanup.txt` removes some now redundant work, and hence gets a mild speedup. If preferred, it can be squashed into the third commit, but like this the algorithmic changes are easier to understand.

```
goos: linux
goarch: arm64
pkg: github.com/cilium/cilium/pkg/ipcache
                      │   fix.txt   │            cleanup.txt             │
                      │   sec/op    │   sec/op     vs base               │
IPCacheUpsert10-10      48.84µ ± 5%   47.59µ ± 5%       ~ (p=0.165 n=10)
IPCacheUpsert100-10     508.3µ ± 5%   485.0µ ± 2%  -4.58% (p=0.005 n=10)
IPCacheUpsert1000-10    5.282m ± 5%   5.056m ± 2%  -4.28% (p=0.011 n=10)
IPCacheUpsert10000-10   48.63m ± 5%   44.87m ± 5%  -7.74% (p=0.009 n=10)
geomean                 1.589m        1.513m       -4.81%
```

Note that simply reverting 93cd67f2994a618865fbcdb0294bcd9b4d44036a is not actually a fix, as the quadratic behaviour still lurks. This can be observed by forcefully setting `ipcache.needNamedPorts = true` when benchmarking:

```
                      │ main-head.txt  │    revert-with-neednamedports.txt     │ 
                      │     sec/op     │     sec/op      vs base               │  
IPCacheUpsert10-10        95.00µ ±  5%     92.12µ ±  3%  -3.04% (p=0.000 n=10)   
IPCacheUpsert100-10      2729.7µ ±  4%    2617.5µ ± 12%       ~ (p=0.165 n=10)   
IPCacheUpsert1000-10     70.192m ± 11%    68.974m ± 10%       ~ (p=0.579 n=10)   
IPCacheUpsert10000-10   4247.52m ±  3%   4233.62m ±  1%       ~ (p=0.853 n=10)   
geomean                   16.68m           16.29m        -2.31%                 
```

Fixes: #24987 

```release-note
Address cilium-agent startup performance regression.
```