@bimmlerd 2023-04-21
> technically the namedport map is not the only lock being held while it's locked; the ipcache lock is always held before the namedport map lock is taken. Actually, it seems like we should "promote" this relationship to say that the ipcache lock must be taken before the namedport map.

What I was trying to express in the commit message was that we don't _acquire_ a new lock while holding the namedport mutex (M), and hence don't establish any relations of the form M > X, for any X. I think that's correct. But yes, we do hold the ipcache mutex while acquiring the namedport mutex, and establish IPCache > M. This is precisely why it's crucial to not acquire any other mutex while also holding M.

Tests were green before pushing to clarify the commitmsg and fix a cosmetic issue in the benchmark, but more reruns can't hurt.