@bimmlerd 2023-04-25
> but if it's at all possible to split refactoring and key changes into separate commits, that would help subsequent review and debugging if we ever have to go back and `git bisect` this change.

@joestringer It took a bit of doing, but I've split the changes up into four commits, see the edited PR description or the commit messages. I hope the changes are a bit clearer now, thanks for the suggestion. PTAL.