@bimmlerd 2023-04-25
Previous CI failures triage:
- Cilium-PR-K8s-1.25-kernel-5.4 failed with https://github.com/cilium/cilium/issues/24966
- Cilium-PR-K8s-1.27-kernel-net-next failed with something I haven't seen before - creating an issue.