@gandro 2023-04-26
Removing the ready-to-merge label till David had time to look at the optional feedback. Feel free to re-add it you don't plan further changes.