@bimmlerd 2023-04-27
CI triage:
- k8s-1.26-kernel-net-next (test-1.26-net-next) failed with https://github.com/cilium/cilium/issues/24687 (we don't have the fix commit 0691784961b76be4cc93f3d7489ec55166a60ebe)

I'll gamble one rerun and otherwise rebase and run it all again